#37, finish making project structure feature-based.
