+++ src/main/java/com/david/caterest/auth/AuthenticationResponse.java
@@ -1,0 +1,15 @@
+package com.david.caterest.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticationResponse { // temp for controller return
+    private Long userId;
+    private String username;
+}
