+++ src/main/java/com/david/caterest/comment/CommentRepository.java
@@ -1,0 +1,6 @@
+package com.david.caterest.comment;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+}
