--- src/main/java/com/david/caterest/dto/AuthenticationResponse.java
@@ -1,15 +1,0 @@
-package com.david.caterest.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationResponse { // temp for controller return
-    private Long userId;
-    private String username;
-}
