--- src/main/java/com/david/caterest/dto/picture/PictureIdDto.java
@@ -1,9 +1,0 @@
-package com.david.caterest.dto.picture;
-
-import lombok.Data;
-
-@Data
-public class PictureIdDto {
-
-    private long id;
-}
