--- src/main/java/com/david/caterest/dto/picture/PicturePostDto.java
@@ -1,8 +1,0 @@
-package com.david.caterest.dto.picture;
-
-import lombok.Data;
-
-@Data
-public class PicturePostDto {
-    String description;
-}
