--- src/main/java/com/david/caterest/dto/user/UserDto.java
@@ -1,10 +1,0 @@
-package com.david.caterest.dto.user;
-
-import lombok.Data;
-
-@Data
-public class UserDto {
-    private Long id;
-    private String displayName;
-
-}
