--- src/main/java/com/david/caterest/dto/user/UserLogInDto.java
@@ -1,13 +1,0 @@
-package com.david.caterest.dto.user;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Data;
-
-@Data
-public class UserLogInDto {
-
-    @NotNull
-    private String email;
-    @NotNull
-    private String password;
-}
