--- src/main/java/com/david/caterest/entity/Role.java
@@ -1,7 +1,0 @@
-package com.david.caterest.entity;
-
-public enum Role {
-
-    USER,
-    ADMIN
-}
