+++ src/main/java/com/david/caterest/picture/PictureRepository.java
@@ -1,0 +1,8 @@
+package com.david.caterest.picture;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
+
+public interface PictureRepository extends JpaRepository<Picture, Long> {
+    public List<Picture> findAllByOrderByDateOfPostDesc();
+}
