+++ src/main/java/com/david/caterest/picture/dto/PictureIdDto.java
@@ -1,0 +1,9 @@
+package com.david.caterest.picture.dto;
+
+import lombok.Data;
+
+@Data
+public class PictureIdDto {
+
+    private long id;
+}
