+++ src/main/java/com/david/caterest/picture/dto/PicturePostDto.java
@@ -1,0 +1,8 @@
+package com.david.caterest.picture.dto;
+
+import lombok.Data;
+
+@Data
+public class PicturePostDto {
+    String description;
+}
