--- src/main/java/com/david/caterest/repository/CommentRepository.java
@@ -1,7 +1,0 @@
-package com.david.caterest.repository;
-
-import com.david.caterest.entity.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-}
