+++ src/main/java/com/david/caterest/user/Role.java
@@ -1,0 +1,7 @@
+package com.david.caterest.user;
+
+public enum Role {
+
+    USER,
+    ADMIN
+}
