+++ src/main/java/com/david/caterest/user/dto/UserDto.java
@@ -1,0 +1,10 @@
+package com.david.caterest.user.dto;
+
+import lombok.Data;
+
+@Data
+public class UserDto {
+    private Long id;
+    private String displayName;
+
+}
