+++ src/main/java/com/david/caterest/user/dto/UserLogInDto.java
@@ -1,0 +1,13 @@
+package com.david.caterest.user.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Data;
+
+@Data
+public class UserLogInDto {
+
+    @NotNull
+    private String email;
+    @NotNull
+    private String password;
+}
